+++ src/main/java/org/braekpo1nt/mctmanager/commands/mct/admin/AdminCommand.java
@@ -1,6 +1,6 @@
 package org.braekpo1nt.mctmanager.commands.mct.admin;
 
-import org.braekpo1nt.mctmanager.commands.commandmanager.CommandManager;
+import org.braekpo1nt.mctmanager.commands.manager.CommandManager;
 import org.braekpo1nt.mctmanager.games.GameManager;
 
 public class AdminCommand extends CommandManager {
