+++ src/main/java/org/braekpo1nt/mctmanager/commands/mct/event/EventCommand.java
@@ -3,9 +3,9 @@ package org.braekpo1nt.mctmanager.commands.mct.event;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextDecoration;
-import org.braekpo1nt.mctmanager.commands.commandmanager.*;
+import org.braekpo1nt.mctmanager.commands.manager.*;
 import org.braekpo1nt.mctmanager.commands.CommandUtils;
-import org.braekpo1nt.mctmanager.commands.commandmanager.commandresult.CommandResult;
+import org.braekpo1nt.mctmanager.commands.manager.commandresult.CommandResult;
 import org.braekpo1nt.mctmanager.commands.mct.event.finalgame.FinalGameCommand;
 import org.braekpo1nt.mctmanager.commands.mct.event.vote.VoteCommand;
 import org.braekpo1nt.mctmanager.games.GameManager;
