+++ src/main/java/org/braekpo1nt/mctmanager/commands/mct/team/score/set/ScoreSetCommand.java
@@ -1,6 +1,6 @@
 package org.braekpo1nt.mctmanager.commands.mct.team.score.set;
 
-import org.braekpo1nt.mctmanager.commands.commandmanager.CommandManager;
+import org.braekpo1nt.mctmanager.commands.manager.CommandManager;
 import org.braekpo1nt.mctmanager.games.GameManager;
 import org.jetbrains.annotations.NotNull;
